backend/requirements.txt
Flask==3.1.1
Flask-Cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.23.2
psycopg[binary]==3.2.9
python-dotenv==1.1.1
